refactor(broker): split db/mod.rs into error/types/sqlite (C29 of PR-3)
Pure mechanical move. db/mod.rs (1786 lines) becomes:

- db/error.rs (~98 lines): DbError enum, impl_coded_debug!, impl CodedError,
  impl From<sqlx::Error>
- db/types.rs (~54 lines): AggregationOrder (pub) + DbOrder / DbBatch /
  DbLockedRequest crate-private row shapes (fields promoted to pub(crate)
  for cross-module access from sqlite.rs)
- db/sqlite.rs (~1609 lines): URL helpers, SqliteDb struct + impl,
  BrokerDb impl, both #[cfg(test)] modules
- db/mod.rs (~120 lines): module decls, pub use re-exports for DbError /
  SqliteDb / broker_sqlite_url_for_chain / AggregationOrder, the BrokerDb
  trait definition, pub type DbObj alias

External paths preserved: crate::db::DbObj / DbError / SqliteDb /
AggregationOrder / BrokerDb / broker_sqlite_url_for_chain all resolve
unchanged. fuzz_db.rs untouched.
